+++ lua/github-theme/extra/iterm.lua
@@ -34,7 +34,6 @@ end
 
 function M.iterm(config)
   config = config or config_module.config
-  config.transform_colors = true
   local colors = require("github-theme.colors").setup(config)
 
   local groups = {