+++ lua/github-theme/extra/kitty.lua
@@ -5,7 +5,6 @@ local M = {}
 
 function M.kitty(config)
   config = config or config_module.config
-  config.transform_colors = true
   local colors = require("github-theme.colors").setup(config)
 
   local kitty = util.template([[