+++ lua/github-theme/extra/tmux.lua
@@ -6,7 +6,6 @@ local M = {}
 ---@param config Config
 function M.tmux(config)
   config = config or config_module.config
-  config.transform_colors = true
   local colors = require("github-theme.colors").setup(config)
 
   local tmux = util.template([[